+++ src/Composer/Command/BuildClearCacheCommand.php
@@ -47,11 +47,11 @@ class BuildClearCacheCommand extends ProcessCommand
         );
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output): int
+    protected function doExecute(InputInterface $input, OutputInterface $output): int
     {
         $output->writeln('<info>Clearing the build cache...</info>');
 
-        $exitCode = parent::execute($input, $output);
+        $exitCode = parent::doExecute($input, $output);
 
         if ($exitCode !== 0) {
             $output->writeln('<error>Unable to clear the build cache</error>');