+++ tests/Composer/Command/AnalyzeCommandTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Ramsey\Test\Dev\Tools\Composer\Command;
 
 use Composer\Console\Application;
-use Composer\EventDispatcher\EventDispatcher;
 use Mockery\MockInterface;
 use Ramsey\Dev\Tools\Composer\Command\AnalyzeCommand;
 use Symfony\Component\Console\Command\Command;
@@ -38,12 +37,6 @@ class AnalyzeCommandTest extends CommandTestCase
         $input = new StringInput('');
         $output = new NullOutput();
 
-        /** @var EventDispatcher & MockInterface $eventDispatcher */
-        $eventDispatcher = $this->mockery(EventDispatcher::class, [
-            'dispatch' => null,
-        ]);
-        $this->getComposer()->setEventDispatcher($eventDispatcher);
-
         /** @var Application & MockInterface $application */
         $application = $this->mockery(Application::class, [
             'getHelperSet' => $this->mockery(HelperSet::class),
@@ -51,15 +44,15 @@ class AnalyzeCommandTest extends CommandTestCase
         $application->shouldReceive('getDefinition')->passthru();
         $application
             ->expects()
-            ->find($this->getCommand()->withPrefix('analyze:phpstan'))
+            ->find($this->command->withPrefix('analyze:phpstan'))
             ->andReturn($commandPhpStan);
         $application
             ->expects()
-            ->find($this->getCommand()->withPrefix('analyze:psalm'))
+            ->find($this->command->withPrefix('analyze:psalm'))
             ->andReturn($commandPsalm);
 
-        $this->getCommand()->setApplication($application);
+        $this->command->setApplication($application);
 
-        $this->assertSame(0, $this->getCommand()->run($input, $output));
+        $this->assertSame(0, $this->command->run($input, $output));
     }
 }