+++ tests/Composer/Command/ProcessCommandTestCase.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Ramsey\Test\Dev\Tools\Composer\Command;
 
-use Composer\EventDispatcher\EventDispatcher;
 use Mockery\MockInterface;
 use Ramsey\Dev\Tools\Composer\Command\ProcessCommand;
 use Ramsey\Dev\Tools\Process\Process;
@@ -36,19 +35,14 @@ abstract class ProcessCommandTestCase extends CommandTestCase
         $this->output = $this->mockery(OutputInterface::class);
     }
 
-    protected function getCommand(): ProcessCommand
-    {
-        /** @var ProcessCommand $command */
-        $command = parent::getCommand();
-
-        return $command;
-    }
-
     public function testGetProcessCommand(): void
     {
+        /** @var ProcessCommand $command */
+        $command = $this->command;
+
         $this->assertSame(
             $this->processCommand,
-            $this->getCommand()->getProcessCommand($this->input, $this->output),
+            $command->getProcessCommand($this->input, $this->output),
         );
     }
 
@@ -59,12 +53,6 @@ abstract class ProcessCommandTestCase extends CommandTestCase
 
     public function doTestRun(callable $processCommandCallbackTest, int $exitCode): void
     {
-        /** @var EventDispatcher & MockInterface $eventDispatcher */
-        $eventDispatcher = $this->mockery(EventDispatcher::class, [
-            'dispatch' => null,
-        ]);
-        $this->getComposer()->setEventDispatcher($eventDispatcher);
-
         /** @var Process & MockInterface $process */
         $process = $this->mockery(Process::class);
         $process->expects()->start();
@@ -73,8 +61,7 @@ abstract class ProcessCommandTestCase extends CommandTestCase
             ->once()
             ->andReturnUsing($processCommandCallbackTest);
 
-        $this
-            ->getProcessFactory()
+        $this->processFactory
             ->expects()
             ->factory($this->processCommand, $this->repositoryRoot)
             ->andReturn($process);
@@ -82,7 +69,7 @@ abstract class ProcessCommandTestCase extends CommandTestCase
         $this->input->shouldIgnoreMissing();
         $this->output->expects()->write('test buffer string');
 
-        $this->assertSame($exitCode, $this->getCommand()->run($this->input, $this->output));
+        $this->assertSame($exitCode, $this->command->run($this->input, $this->output));
     }
 
     private function getProcessCommandCallbackTest(): callable